Zoom following mouse

+++ libft/src/ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: pvital-m <pvital-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/23 18:03:25 by pvital-m          #+#    #+#             */
-/*   Updated: 2023/03/10 22:59:45 by pvital-m         ###   ########.fr       */
+/*   Updated: 2023/06/09 13:53:14 by pvital-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	long long	i;
+	size_t	i;
 	char		*d_dst;
 	char		*s_src;
